Create app dir to contain everything but the bins
