+++ src/app/cakes/cake.rs
@@ -1,18 +1,22 @@
 use sea_orm::entity::prelude::*;
-use crate::lil_lib::bakeries as bakeries;
-use crate::lil_lib::lineitems as lineitems;
-use crate::lil_lib::customers as customers;
+use rocket::serde::{Serialize, Deserialize};
+use crate::app::bakeries as bakeries;
+use crate::app::bakers as bakers;
+use crate::app::cakes_bakers as cakes_bakers;
+use crate::app::lineitems as lineitems;
 
-#[derive(Clone, Debug, PartialEq, DeriveEntityModel)]
-#[sea_orm(table_name = "order")]
+#[derive(Clone, Debug, PartialEq, DeriveEntityModel, Serialize, Deserialize)]
+#[serde(crate = "rocket::serde")]
+#[sea_orm(table_name = "cake")]
 pub struct Model {
     #[sea_orm(primary_key)]
     pub id: i32,
+    pub name: String,
     #[sea_orm(column_type = "Decimal(Some((19, 4)))")]
-    pub total: Decimal,
-    pub bakery_id: i32,
-    pub customer_id: i32,
-    pub placed_at: DateTime,
+    pub price: Decimal,
+    pub bakery_id: Option<i32>,
+    pub gluten_free: bool,
+    pub serial: Uuid,
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
@@ -25,14 +29,6 @@ pub enum Relation {
         on_delete = "Cascade"
     )]
     Bakery,
-    #[sea_orm(
-        belongs_to = "customers::customer::Entity",
-        from = "Column::CustomerId",
-        to = "customers::customer::Column::Id",
-        on_update = "Cascade",
-        on_delete = "Cascade"
-    )]
-    Customer,
     #[sea_orm(has_many = "lineitems::lineitem::Entity")]
     Lineitem,
 }
@@ -43,9 +39,13 @@ impl Related<bakeries::bakery::Entity> for Entity {
     }
 }
 
-impl Related<customers::customer::Entity> for Entity {
+impl Related<bakers::baker::Entity> for Entity {
     fn to() -> RelationDef {
-        Relation::Customer.def()
+        cakes_bakers::Relation::Baker.def()
+    }
+
+    fn via() -> Option<RelationDef> {
+        Some(cakes_bakers::Relation::Cake.def().rev())
     }
 }
