+++ src/app/cakes/handler.rs
@@ -2,12 +2,13 @@ use rocket::http::Status;
 use rocket::serde::json::{json, Json, Value};
 use rocket_db_pools::{Connection};
 use sea_orm::{entity::*, query::*};
-use crate::db::pool::Db;
-// use crate::lil_lib::bakery_chain::*;
+// use crate::app::bakery_chain::*;
 use super::cake::Entity as Cake;
 
+use super::pool;
+
 #[get("/")]
-pub async fn all(connection: Connection<Db>) -> Result<Json<Vec<super::cake::Model>>, Status> {
+pub async fn all(connection: Connection<pool::Db>) -> Result<Json<Vec<super::cake::Model>>, Status> {
       Ok(Json(Cake::find()
         .all(&connection)
         .await
