+++ src/app/cakes/mod.rs
@@ -1,2 +1,3 @@
 pub mod handler;
 pub mod cake;
+pub use super::pool;
