+++ src/app/customers/customer.rs
@@ -1,5 +1,5 @@
 use sea_orm::entity::prelude::*;
-use crate::lil_lib::orders as orders;
+use crate::app::orders as orders;
 
 #[derive(Clone, Debug, PartialEq, DeriveEntityModel)]
 #[sea_orm(table_name = "customer")]
