+++ src/app/lineitems/lineitem.rs
@@ -1,6 +1,6 @@
 use sea_orm::entity::prelude::*;
-use crate::lil_lib::cakes as cakes;
-use crate::lil_lib::orders as orders;
+use crate::app::cakes as cakes;
+use crate::app::orders as orders;
 
 #[derive(Clone, Debug, PartialEq, DeriveEntityModel)]
 #[sea_orm(table_name = "lineitem")]
