+++ src/main.rs
@@ -5,9 +5,8 @@ use rocket::serde::json::{json, Value};
 use rocket::{Build, Rocket};
 use rocket_db_pools::{Database};
 
-mod lil_lib;
-mod db;
-use db::{pool, migrations};
+mod app;
+use app::db::{pool, migrations};
 
 async fn run_migrations(rocket: Rocket<Build>) -> fairing::Result {
     let conn = &pool::Db::fetch(&rocket).unwrap().conn;
@@ -29,11 +28,11 @@ fn rocket() -> _ {
         .attach(pool::Db::init())
         .attach(AdHoc::try_on_ignite("Migrations", run_migrations))
         .mount("/cakes",
-            routes![lil_lib::cakes::handler::all,
-                // lil_lib::cakes::handler::get,
-                // lil_lib::cakes::handler::post,
-                // lil_lib::cakes::handler::put,
-                // lil_lib::cakes::handler::delete
+            routes![app::cakes::handler::all,
+                // app::cakes::handler::get,
+                // app::cakes::handler::post,
+                // app::cakes::handler::put,
+                // app::cakes::handler::delete
                 ],
         )
         .register("/", catchers![not_found])
